clarified some texture binding and loading (hopefully)
Managing multiple texture units can be one of the most obtuse things in opengl, and is often not even necessary.

I tried to demonstrate the use of 2 texture units, but it's not entirely necessary in this simple case.  You're more likely just to use a single texture unit and "set it and forget it" then just change which texture handle is bound to the default texture unit